Move all storage from HTML5 API to chrome API
Removes all window.storage references from the project.
Now all data are stored as objects instead of strings so there's no need for
serialization/deserializarion.
Also using chrome.storage enables the popup and the content script to retrieve
the data they need natively without any need for message passing. All logic concerning
message passing is removed. This simplifies the logic of the extension.

The extension now supports data sync across devices using chrome.storage.sync.
The user will be able to opt in from the options page. Needless to say that, this
works only if user has enable sync in his browser.

A lot of functions with callbacks have been added so as to facilitate the
transition. Some of them  may be revised in the future.